Major rewrite of sds_tool!
This change makes it _much_ easier to add logic to decode messages.
Just add an entry to a data struct and optionally write a function
which can return the result as a string.  Allows us to even decode
a single value multiple ways with multiple formulas.

The tool is still "passive" meaning it doesn't do a K-Line init
so it still requires a SDS Tool to do all the talking.