add sds_print.ino and read_serial.py
Realized that doing all the decoding in the Teensy was too painful.
Using sds_print to handle the framing and only printing the messages
from the ECU we are able to the decode them using read_serial.py

Currently sds_print does not query the ECU so you will need an SDS Tool
to do so.